requirements.txt
matplotlib==3.5.1
natsort==8.0.2
numpy==1.22.1
pandas==1.4.0
Pillow==9.0.0
streamlit==1.4.0
tensorflow==2.7.0
tqdm==4.62.3
